+++ CHANGELOG.md
@@ -9,6 +9,8 @@
 
 ## Unreleased
 
+## 16.27.1
+
 * Add descriptions for search actions on machine readable components (PR #897)
 
 ## 16.27.0